// book-network-ui/src/app/services/models/api-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ApiResponse {
  Date?: any;
  ListData?: Array<any>;
  Message?: string;
  StatusCode?: number;
  accessToken?: string;
  refreshToken?: string;
  role?: string;
  userId?: number;

  error?: string;
  businessErrorCode?: number;
  businessErrorDescription?: string;
}